¨📝¨

+++ proyecto1/src/proyecto1/Word.java
@@ -3,7 +3,7 @@ package proyecto1;
 public class Word {
     String word;
     boolean found = false;
-    float time = 0;
+    long time = 0;
     
     public Word(String word){
         this.word = word;
